Add files via upload
A little game prototype I worked on during my early years of college, it's a simple bullet-hell Arkanoid clone akin to "Touhou Highly Responsive to Prayers" where I tried to introduce some more mechanics into the formula.
The physics for the main sphere you need to bounce up to hit the bricks is a bit wonky, never got time to finish the prototype fully; might work on it a bit more in the future as a side project. After the boss battle you are instantly put into a debug level where I tested the mechanics I added into the game.
The game has a "level-skip debug button"; just press ENTER whenever in a level to skip it (yes that includes the boss battle after the 5th level).